Apple Pie (#8)
* :beer: substantially update project structure


* :aerial_tramway: Racecarmode!

* :cake: bell
