+++ chat-app/.prettierrc
@@ -1,0 +1,5 @@
+{
+    "tabWidth": 4,
+    "semi": true,
+    "singleQuote": false
+}
