+++ chat-app/src/components/panels/index.js
