+++ chat-app/src/pages/main/index.js
@@ -19,8 +19,8 @@ import {
     Select,
     MenuItem,
 } from "@material-ui/core";
-import MessageBox from "./components/message/Message";
-import Panels from "./components/panels/Panels";
+import MessageBox from "../../components/message/index";
+import Panels from "../../components/panels/index";
 import axios from "axios";
 import { useTranslation } from "react-i18next";
 
@@ -190,7 +190,7 @@ export default function Chat(props) {
             });
     };
 
-    useEffect(setChatrooms, []); // set chatrooms after entering the chat page
+    useEffect(setChatrooms, [username]); // set chatrooms after entering the chat page
 
     useEffect(() => {
         if (rooms === undefined || rooms.length === 0) {
@@ -301,22 +301,9 @@ export default function Chat(props) {
     };
 
     async function sendChatMessage() {
-        let UnixTime = await getTime();
-        sendMessage(UnixTime);
+        let response = await axios.get(REACT_APP_SERVER_ADDRESS + "/time");
+        sendMessage(response.data.UTCTime.UnixTime);
     }
-
-    const getTime = () => {
-        return new Promise((resolve, reject) => {
-            fetch(REACT_APP_SERVER_ADDRESS + "/time", {
-                method: "GET",
-            }).then((response) => {
-                response.json().then((data) => {
-                    let time = data.UTCTime.UnixTime;
-                    resolve(time);
-                });
-            });
-        });
-    };
 
     const sendMessage = (time) => {
         setChatText("");
