+++ chat-app/src/pages/sign-up/index.js
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import Avatar from "@material-ui/core/Avatar";
 import Button from "@material-ui/core/Button";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import TextField from "@material-ui/core/TextField";
@@ -7,7 +8,9 @@ import Grid from "@material-ui/core/Grid";
 import Box from "@material-ui/core/Box";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
+import AcUnitIcon from "@material-ui/icons/AcUnit";
 import { useTranslation } from "react-i18next";
+import { signInOrOut } from "../../utils/HttpRequest";
 
 function Copyright() {
     // i18n
@@ -45,7 +48,7 @@ const useStyles = makeStyles((theme) => ({
     },
     form: {
         width: "100%", // Fix IE 11 issue.
-        marginTop: theme.spacing(1),
+        marginTop: theme.spacing(3),
     },
     submit: {
         margin: theme.spacing(3, 0, 2),
@@ -55,77 +58,73 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-const { REACT_APP_SERVER_ADDRESS } = process.env;
-
-export default function SignIn(props) {
+export default function SignUp(props) {
     const classes = useStyles();
 
     // i18n
     const { t } = useTranslation();
 
     const setPage = props.setPage;
-    const setUser = props.setUser;
     const [username, setUsername] = useState("");
     const [password, setPassword] = useState("");
 
-    const onSubmit = (e) => {
+    const onSubmit = async (e) => {
         e.preventDefault();
-        let formData = new FormData();
-        formData.append("username", username);
-        formData.append("password", password);
-        fetch(REACT_APP_SERVER_ADDRESS + "/users/login", {
-            method: "POST",
-            body: formData,
-        }).then((response) => {
+        try {
+            let response = await signInOrOut("/users/register", {
+                username: username,
+                password: password,
+            });
             if (response.status === 200) {
-                response.json().then((data) => {
-                    let user = {
-                        name: "",
-                    };
-                    user.name = data.username;
-                    setUser(user);
-                    localStorage.setItem("token", data.token);
-                });
+                setPage("sign-in");
             }
-        });
-        setUsername("");
-        setPassword("");
+            setUsername("");
+            setPassword("");
+        } catch (error) {
+            console.log(error);
+        }
     };
 
     return (
         <div className={classes.outerContainer}>
             <CssBaseline />
             <div className={classes.paper}>
+                <Avatar className={classes.avatar}>
+                    <AcUnitIcon />
+                </Avatar>
                 <Typography component="h1" variant="h3">
-                    Chat!
+                    {t("signUp.title")}
                 </Typography>
                 <form className={classes.form} noValidate onSubmit={onSubmit}>
-                    <TextField
-                        variant="outlined"
-                        margin="normal"
-                        required
-                        fullWidth
-                        id="Username"
-                        label={t("signIn.username")}
-                        name="username"
-                        autoComplete="username"
-                        autoFocus
-                        value={username}
-                        onChange={(e) => setUsername(e.target.value)}
-                    />
-                    <TextField
-                        variant="outlined"
-                        margin="normal"
-                        required
-                        fullWidth
-                        name="password"
-                        label={t("signIn.password")}
-                        type="password"
-                        id="password"
-                        autoComplete="current-password"
-                        value={password}
-                        onChange={(e) => setPassword(e.target.value)}
-                    />
+                    <Grid container spacing={2}>
+                        <Grid item xs={12}>
+                            <TextField
+                                variant="outlined"
+                                required
+                                fullWidth
+                                id="username"
+                                label={t("signUp.nickname")}
+                                name="username"
+                                autoComplete="username"
+                                value={username}
+                                onChange={(e) => setUsername(e.target.value)}
+                            />
+                        </Grid>
+                        <Grid item xs={12}>
+                            <TextField
+                                variant="outlined"
+                                required
+                                fullWidth
+                                name="password"
+                                label={t("signUp.password")}
+                                type="password"
+                                id="password"
+                                autoComplete="current-password"
+                                value={password}
+                                onChange={(e) => setPassword(e.target.value)}
+                            />
+                        </Grid>
+                    </Grid>
                     <Button
                         type="submit"
                         fullWidth
@@ -133,23 +132,18 @@ export default function SignIn(props) {
                         color="primary"
                         className={classes.submit}
                     >
-                        {t("signIn.signInButton")}
+                        {t("signUp.signUpButton")}
                     </Button>
-                    <Grid container>
-                        <Grid item xs>
-                            <Link href="#" variant="body2">
-                                Forgot password?
-                            </Link>
-                        </Grid>
+                    <Grid container justify="flex-end">
                         <Grid item>
                             <Link
                                 href="#"
                                 variant="body2"
                                 onClick={() => {
-                                    setPage("sign-up");
+                                    setPage("sign-in");
                                 }}
                             >
-                                {t("signIn.goToSignUp")}
+                                {t("signUp.goToSignIn")}
                             </Link>
                         </Grid>
                     </Grid>
